Add option to preserve or not the Z position
Objects can now preserve or not their position in the Z axis, if the
position is preserved then when the "magnified" object is added to the
global file, the object will be positioned exactly where the
"non-magnified" object was. In any other case the object will simply
retain the (x, y) positioning and will be drawn at the very end of the
global file, making it such that it will appear infront of any of the
other objects that were in the scene.