Most recent notebook for filtering
This will likely change soon as we find a better way to find the threshold
that is used to identify contaminant sequences in an OTU.  Currently, the
threshold is set to 90%; if a single sequence composes >=90% of the over-
represented gammaproteo OTU, then the sequence is labeled as a contaminant.
Instead, we would like to define this threshold based on the sequencing
error rate as determined by the Illumina instrument.